Move the collect() function to iutil
The collect() function is no longer used only by UI code,
so move it out of the UI namespace. This should also reduce
the likelihood of import loops.
